Create principal_component_analysis_for_image_fusion.m
Principal components analysis (PCA) is one of a family of techniques for
taking high-dimensional data, and using the dependencies between the variables to represent it in a more tractable, lower-dimensional form, without losing too much information. PCA is one of the simplest and most robust ways of doing such dimensionality reduction. 
The above algorithm does an orthogonal transformation to convert a set of observations of possibly correlated variables into a set of values of linearly uncorrelated variables called principal components.
The above algorithm is used to fuse 2 images.